fix: 이메일과 비밀번호를 받아서 회원을 검증하는 메서드명 수정
- docs: 구조 변경에서 변경되지 않은 테스트 코드 및 시나리오 수정

- fix: Member Controller import 정리
