+++ src/main/java/nextstep/core/auth/application/UserDetailsService.java
@@ -2,7 +2,7 @@ package nextstep.core.auth.application;
 
 public interface UserDetailsService {
 
-    boolean validateUser(String email, String password);
+    boolean verifyUser(String email, String password);
 
     String findOrCreate(String email);
 }
