+++ src/test/java/nextstep/core/member/application/UserDetailsServiceImplTest.java
@@ -1,9 +1,6 @@
-package nextstep.core.auth.application;
+package nextstep.core.member.application;
 
 import nextstep.common.annotation.ApplicationTest;
-import nextstep.core.auth.application.dto.TokenResponse;
-import nextstep.core.member.application.MemberService;
-import nextstep.core.member.application.UserDetailsServiceImpl;
 import nextstep.core.member.application.dto.MemberRequest;
 import nextstep.core.member.exception.NotFoundMemberException;
 import org.junit.jupiter.api.BeforeEach;
@@ -18,7 +15,7 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 @DisplayName("유저 서비스 레이어 테스트")
 @ApplicationTest
-public class UserDetailServiceTest {
+public class UserDetailsServiceImplTest {
 
     UserDetailsServiceImpl userDetailsService;
 
@@ -37,17 +34,17 @@ public class UserDetailServiceTest {
         class 성공 {
             /**
              * Given 회원을 생성한다.
-             * When  생성된 회원 정보로 토큰 발급을 요청할 경우
-             * Then  정상적으로 토큰이 발급된다.
+             * When  해당 회원의 이메일과 비밀번호를 확인할 경우
+             * Then  정상적으로 확인된다.
              */
             @Test
-            void 저장된_회원으로_토큰_발급_요청() {
+            void 저장된_회원으로_확인_요청() {
                 // given
                 MemberRequest memberRequest = new MemberRequest(브라운.이메일, 브라운.비밀번호, 브라운.나이);
                 memberService.createMember(memberRequest);
 
                 // when
-                boolean actual = userDetailsService.validateUser(브라운.이메일, 브라운.비밀번호);
+                boolean actual = userDetailsService.verifyUser(브라운.이메일, 브라운.비밀번호);
 
                 // then
                 assertThat(actual).isTrue();
@@ -58,11 +55,11 @@ public class UserDetailServiceTest {
         class 실패 {
             /**
              * Given 회원을 생성한다.
-             * When  생성된 회원의 비밀번호가 아닌 회원 정보로 토큰 발급을 요청할 경우
-             * Then  토큰이 발급되지 않는다.
+             * When  해당 회원의 비밀번호가 아닌 다른 비밀번호로 확인할 경우
+             * Then  확인할 수 없다.
              */
             @Test
-            void 비밀번호가_다른_회원의_토큰_발급_요청() {
+            void 비밀번호가_다른_회원의_확인_요청() {
                 // given
                 String changedPassword = "changed password";
 
@@ -72,21 +69,21 @@ public class UserDetailServiceTest {
                 // when, then
                 assertThatExceptionOfType(IllegalArgumentException.class)
                         .isThrownBy(() -> {
-                            userDetailsService.validateUser(스미스.이메일, changedPassword);
+                            userDetailsService.verifyUser(스미스.이메일, changedPassword);
                         }).withMessageMatching("비밀번호가 다릅니다.");
             }
 
             /**
              * Given 저장소에 저장하지 않는 회원을 생성한다.
-             * When  회원 저장소에 존재하지 않는 회원 정보로 토큰 발급을 요청할 경우
-             * Then  토큰이 발급되지 않는다.
+             * When  회원 저장소에 존재하지 않는 회원 정보를 확인할 경우
+             * Then  확인할 수 없다.
              */
             @Test
-            void 존재하지_않는_회원의_토큰_발급_요청() {
+            void 존재하지_않는_회원의_확인_요청() {
                 // when, then
                 assertThatExceptionOfType(NotFoundMemberException.class)
                         .isThrownBy(() -> {
-                            userDetailsService.validateUser(존슨.이메일, 존슨.비밀번호);
+                            userDetailsService.verifyUser(존슨.이메일, 존슨.비밀번호);
                         }).withMessageMatching("회원 정보가 없습니다.");
             }
